# backend/internal/model/slam.go
package model

type Slam struct {
	Id          int    `json:"id"`
	Title       string `json:"title" validate:"required"`
	Description string `json:"description"`
	Public      bool   `json:"public"`
}